rename blaze module to shard
